# crc16: Accelerate T10DIF performance with prefetch and pmull2