The memory block size calculated by t10dif is generally 512 bytes in
sectors. prefetching can effectively reduce cache misses.Use ldp instead
of ldr to reduce the number of instructions, pmull+pmull2 can resuce
register access. The perf test result shows that the performance is
improved by 5x ~ 14x after optimization.

The crc16_t10pi_op_copy_insert_cold test performance is improved from 229.75 MB/s to 3399.94 MB/s,
The performance in the crc16_t10dif_perf test is improved from 1493.37 MB/s to 8747.10 MB/s. The performance in the crc16_t10dif_copy_perf test is improved from 1392.13 MB/s to 7512.42 MB/s.

Signed-off-by: Chunsong Feng <fengchunsong@huawei.com>